django_root/requirements.txt
celery
Django
django-cors-headers
django-extensions
django-environ
django-nested-admin
django-phonenumber-field
django-soft-delete
djangorestframework
djangorestframework-simplejwt
django-filter
drf-yasg
drfasyncview
gunicorn
mysqlclient
psycopg2
phonenumberslite
Pillow
PyJWT
pytest
python-dateutil
spacy
redis
traitlets
uvicorn